Moving cutouts endpoint from queries to surveys
