+++ src/api/routes/mod.rs
@@ -5,4 +5,5 @@ pub mod filters;
 pub mod info;
 pub mod kafka;
 pub mod queries;
+pub mod surveys;
 pub mod users;
