+++ src/api/routes/queries/mod.rs
@@ -1,11 +1,9 @@
 pub mod cone_search;
 pub mod count;
-pub mod cutouts;
 pub mod find;
 pub mod pipeline;
 
 pub use cone_search::post_cone_search_query;
 pub use count::{post_count_query, post_estimated_count_query};
-pub use cutouts::get_cutouts;
 pub use find::post_find_query;
 pub use pipeline::post_pipeline_query;
