+++ src/api/routes/surveys/mod.rs
@@ -1,0 +1,3 @@
+pub mod cutouts;
+
+pub use cutouts::get_cutouts;
